[New Groups] Realistic Water & Enhanced Lighting
Both will load after late loaders, just to be sure that they win against any thing that could be placed in Late Loaders.

ELFX - Hardcore.esp has not been added to the enhanced lighting group because of its conflict with Alt start, if the starting room bounds are fixed in ELFX - Hardcore.esp, it should be placed in the eleGroup.