Brotherhood of Old - Add Skyrim Sewers 3rd Party Patch Message
Problem:
Skyrim Sewers and The Brotherhood of Old place their entry doors inside the same Markarth Warrens' room which making The Brotherhood of Old's trap door inaccessible.

Solution
I moved the entry door and the exit point of Skyrim sewer to another unused room of the Markarth Warrens.

This patch can be merged with everything