Distinct Interiors: Properly handle the deleted navmeshes
As it turns out, the one deleted navmesh in this mod (in both the All In One version and the Modular version) is actually a wild edit.
This adds a *hasWildEdits warning with instructions on how to remove it and marks the mods clean once the wild edits have been removed.