Thunderchild - Mark as do not clean.
- Added until Author has time to test and confirm removal of ITMs have no side effect.

Discussion can be found on SSE Thunderchild posts.

Quote from author

"The location reference is nonexistent in the vanilla game but Thunderchild adds it, apparently as a result of updating the High Hrothgar navmesh. Will the navmesh still work without them? Maybe, maybe not, but I assume the CK added them for a reason. The last time I deleted a "useless ITM" from Thunderchild, it resulted in wolf and bear spawns migrating to High Hrothgar, and experimenting with a mod that isn't in a hurry to be updated anyway could end up being a timesink with uncertain benefits.

I added this to the list of things to do for Thunderchild, though. When the list gets big enough to warrant an update, then I'll play around with it."