PCE - Relighting Skyrim.esp: Add C.Water tag
This is a clever way to work around the fact that PCE does not forward water changes from Update.esm.
This patch does not include any of those edits, but it does include the 'Has Water' changes we would lose out on without C.Water.
So, adding the tag here allows RLS users to have the 'Has Water' flags without introducing any of the problematic water changes.