Correct Aliases used for Patch Messages
Alias hasPluginPatch will be used when patches are provided on the mods
page in a seperate download.
The alias haspatchIncluded will be used for patches included with the
plugins installer.

Seperate Obsidian 1st & 3rd party patch links:
These should be kept separate to provide the user with both options,
as they have different approaches to patching these mods.
Both will still only show if the other is not present.

Change True Storm Merged Compatibility patch message to 3rd party:
The Merged Compatibility patches are provided on a seperate page by a
third party and are not supported by the main authors.