Refactor Cleaning Messages
Remove Clean entries with deleted nav:
When a DLC is still dirty (e.g it has deleted nav) it will be left blank, the same is done for Fallout 4, and I think it's important to be consistent, especially with cleaning.

Reworded QAC Dirty messages:
Requires manual fix now includes a link with more information on deleted navmeshes.
Shortened Quickclean message, any information on QAC can be provided on the linked guide page.

Seperate Wild edit message and cleaning guide link:
I've seperated these as sometimes a mod has wild edits, but no cleaning guide is available.
This way LOOT can still inform the user of possible issues, even if a fix is not provided.
I think it's a better approach to always provide a warning if wild edits are present.
And to provide a linked cleaning guide if available.
Remove Built in cleaning guides:
These in built cleaning messages feel very clunky and seem to provide more confusion than help.
The webpage guides we have links to provide better formatting, which are much easier to follow.