Replace hasPluginPatch with patchProvided
To be used in situations where a patch is provided on the plugins mod page.
Removed sub link to plugins own mod page as it seems redundant when a link is already provided in locations.
The name of and optionally a link to the Conflicting mod will still be provided.