Move Bijin mods to default Group
On review, the idea of this group was to allow mods to load after higher groups if nothing was blocking it. Blocking would occur from mods in default by load after rules or dependent modules. This would mean the user would have to add these blocking dependent modules to the addons group to restore previous sorting. It can appear inconsistent if you're unaware of how this works. And it may not be as intuative as I had hoped. Another solution for loading mods after the ICAIO -> RDO block of mods is required.