iNeed Extended - Move to underrides group
Moved early due to the number of cells it places objects inside. This will conflict with any mods that change Cells or alter landscape or placed snowdrifts. It will also conflict with any mods that modify food. A patch is required in any case so I've placed this early to avoid it reverting changes made by other mods.