Ducks and Swans - Move To Pre-Early
Mod page Notes

Compatibility:

Of course there are lots of cells touched by this mod, but as this concerns only water surfaces, this should not be a big issue.
I strongly recommend to put this mod at the beginning of your load order and to load all your other mods after "ducks and swans".