iNeed - Add to Underrides
Moved early due to the number of cells it places objects inside. This will conflict with any mods that change Cells or alter landscape, in particular placed snowdrifts. It will also conflict with any mods that modify food. A patch is usually required in any case, so I've placed this before default to avoid it reverting changes made by other mods.