Winterhold Restored - Sort After Soulmancer Music Merge
SMM removes two CELL flags `hand changed`, I'm not sure what these do exactly.
But for the sake of caution, I think it would be better to let WR win.
Safe to sort after SMM, It has no conflicts with any of the mods
that SMM is set to sort after.