Allow RS Children ESM to be sorted after USSEP
I've tested this order and have no issue with getting to the main menu.
As long as the unresolved reference is fixed.
By using the RSC_USSEP_AventusFix.esp or by removing the
unresolved package from RSChildren.esp.
If I had to guess as to why users are still having issues, there
must be something else in their load order that is referencing that
deleted package.