Add main plugin to Skills & Perks group
Because it needs to be before Sacrosanct.
Order with Alternate Start can be either way.
A patch is required if it's before, none if Requiem is after ASLAL.
Having Requiem in the group before ASLAL will allow users to add an
after to Requiem for ASLAL without causing a sorting error.