Add SkyVRaan Shimmering VR Waters load order
According to the author of the mod, cubemap enabler should be at
the very end of load order to overwrite water type data. However,
it could be before or after DynDOLOD depending on how DynDOLOD.esp
was built.

Water grass addon should be loaded after Landscape Fixes for grass
mod.