Update masterlist.yaml
opusGlass' [Falkreath Hauntings](https://www.nexusmods.com/skyrimspecialedition/mods/20733/) contains 3 ITMs. These are required in order to for the Lady Ilinalta encounter to work properly. They are edits to flow data so that the freezing effect that the boss battle happens as is scripted. Water mods and any other mod that edits flow data in that location will cause the water not to freeze and render the boss battle impossible to complete.