Update Wheels of Lull
Removed Wheels of Lull's Bash tags since they're already included in the plugin description. Updated to latest CRC. Added a writing patch someone made for Wheels of Lull to the masterlist that is now incompatible because it edits old records, and we have meticulously fixed typos so it is probably obsolete as well. Author has not responded to me.

I will add a version number to Lull's description when we next update it for more stable conditions.