Account for Community Shaders in dependencies
Certain mods that previously required ENBSeries can now be used
with Community Shaders and its addons. However, LOOT will show
an error if it does not detect ENBSeries.

This commit makes the masterlist check for the presence of
the Community Shaders plugin and the required addons where
applicable in addition to ENBSeries.